fix error when hitting kill button with no current app

+++ shb-plus.js
@@ -92,10 +92,12 @@
     kill.textContent = '☠';
     kill.addEventListener('touchstart', function() {
       var oldApp = window.wrappedJSObject.StackManager.getCurrent();
-      window.wrappedJSObject.SheetsTransition.begin('ltr');
-      window.wrappedJSObject.SheetsTransition.snapLeft(1);
-      window.wrappedJSObject.StackManager.goPrev();
-      oldApp.kill();
+      if (oldApp) {
+        window.wrappedJSObject.SheetsTransition.begin('ltr');
+        window.wrappedJSObject.SheetsTransition.snapLeft(1);
+        window.wrappedJSObject.StackManager.goPrev();
+        oldApp.kill();
+      }
       if('vibrate' in navigator) {
         // ... vibrate for a second
         navigator.vibrate(50);
